# Recursion 0.5

+++ 0x08-recursion/6-is_prime_number.c
@@ -11,7 +11,7 @@
 
 int is_prime_number(int n)
 {
-	int p = sqrt(n);
+	int p = (int) sqrt(n);
 
 	if (n == 0 || n == 1)
 		return (0);
